+++ Pod/Core/TBActorProxyFuture.h
@@ -1,0 +1,13 @@
+//
+//  TBActorProxyFuture.h
+//  ActorKit
+//
+//  Created by Julian Krumow on 07.08.15.
+//  Copyright (c) 2015 Julian Krumow. All rights reserved.
+//
+
+#import "TBActorProxy.h"
+
+@interface TBActorProxyFuture : TBActorProxy
+
+@end